Diet planner/backend/requirements.txt
Flask==2.3.0
Flask-CORS==4.0.0
openai==0.27.0
python-dotenv==1.0.0
Werkzeug==2.3.0
